requirements.txt
Django==1.10
Jinja2==2.9.6
MarkupSafe==1.0
PyYAML==3.12
arrow==0.10.0
binaryornot==0.4.4
chardet==3.0.4
click==6.7
cookiecutter==1.5.1
diff-match-patch==20121119
django-crispy-forms==1.7.0
django-filter==1.1.0
django-fsm==2.6.0
django-import-export==0.5.1
django-money==0.12
django-reversion==2.0.10
django-tables2==1.14.2
et-xmlfile==1.0.1
future==0.16.0
jdcal==1.3
jinja2-time==0.2.0
mock==2.0.0
odfpy==1.3.5
openpyxl==2.4.9
pbr==3.1.1
poyo==0.4.1
py-moneyed==0.7.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
tablib==0.12.1
termcolor==1.1.0
unicodecsv==0.14.1
urwid==1.3.1
whichcraft==0.4.1
xlrd==1.1.0
xlwt==1.3.0
